i117960 Basic: Line Input doesn't work in single-line If
i92642 added the ability to use certain keywords as variable names (eg. name = 1, line = "hi"),
but also caused a regression where "Line Input" is broken in single-line If statements.
This patch fixes that by allowing Then and Else to also be the start-of-line tokens expected to
immediately preceed the "Line" token in order for that "Line" token to be recognized a keyword instead
of a variable name. Also added FVT spreadsheet tests for "Line" as both a variable name and as "Line Input".

Patch by: me

git-svn-id: https://svn.apache.org/repos/asf/openoffice/trunk@1716234 13f79535-47bb-0310-9956-ffa450edef68
(cherry picked from commit 725d867363770a4c47c4b3c6dbe1f359c296604a)

 Author:    Damjan Jovanovic <damjan@apache.org>
 Date:      Tue Nov 24 18:56:10 2015 +0000
 Changes to be committed:
	modified:   main/basic/source/comp/token.cxx
	new file:   test/testuno/data/uno/sc/fvt/Basic Line as variable and Line Input.ods
	modified:   test/testuno/source/fvt/uno/sc/formula/TestFormulaDocs.java